rename gecko terminal chart
