+++ packages/commonwealth/client/scripts/views/components/GeckoTerminalChart/GeckoTerminalChart.scss
@@ -1,9 +1,9 @@
-.GekoChart {
+.GeckoChart {
   width: 650px;
   height: 500px;
 }
 
-.GekoChartContestPage {
+.GeckoChartContestPage {
   height: 500px;
   width: 100%;
 }
