+++ packages/commonwealth/client/scripts/views/components/GeckoTerminalChart/index.ts
@@ -1,0 +1,1 @@
+export * from './GeckoTerminalChart';
