--- packages/commonwealth/client/scripts/views/components/GekoTerminalChart/index.ts
@@ -1,1 +1,0 @@
-export * from './GekoTerminalChart';
